feat(sampling): Reduce the number of /tags requests
